implement basic fstring substitution
