+++ lib/input/src/print.cc
@@ -777,130 +777,6 @@ template <class T> void print_op(T &out, BinaryOperator op) {
     }
 }
 
-auto to_string(FormatSpec::Align x) -> std::string_view {
-    switch (x) {
-        case FormatSpec::Align::left: {
-            return "<";
-        }
-        case FormatSpec::Align::right: {
-            return ">";
-        }
-        case FormatSpec::Align::center: {
-            return "^";
-        }
-        case FormatSpec::Align::number: {
-            return "=";
-        }
-        default: {
-            return "";
-        }
-    }
-}
-
-auto to_string(FormatSpec::Type x) -> std::string_view {
-    switch (x) {
-        case FormatSpec::Type::character: {
-            return "c";
-        }
-        case FormatSpec::Type::binary: {
-            return "b";
-        }
-        case FormatSpec::Type::octal: {
-            return "o";
-        }
-        case FormatSpec::Type::decimal: {
-            return "d";
-        }
-        case FormatSpec::Type::hex_lower: {
-            return "x";
-        }
-        case FormatSpec::Type::hex_upper: {
-            return "X";
-        }
-        case FormatSpec::Type::locale: {
-            return "n";
-        }
-        default: {
-            return "";
-        }
-    }
-}
-
-auto to_string(FormatSpec::Sign x) -> std::string_view {
-    switch (x) {
-        case FormatSpec::Sign::plus: {
-            return "+";
-        }
-        case FormatSpec::Sign::minus: {
-            return "-";
-        }
-        case FormatSpec::Sign::space: {
-            return " ";
-        }
-        default: {
-            return "";
-        }
-    }
-}
-
-auto to_string(FormatSpec::Grouping x) -> std::string_view {
-    switch (x) {
-        case FormatSpec::Grouping::comma: {
-            return ",";
-        }
-        case FormatSpec::Grouping::underscore: {
-            return "_";
-        }
-        default: {
-            return "";
-        }
-    }
-}
-
-auto to_string(FormatSpec::Conversion x) -> std::string_view {
-    switch (x) {
-        case FormatSpec::Conversion::repr: {
-            return "!r";
-        }
-        default: {
-            return "";
-        }
-    }
-}
-
-template <class T> auto print_spec(T &out, FormatSpec const &spec) -> T & {
-    for (auto const &x : spec.accessors) {
-        std::visit(
-            [&out]<class U>(U const &acc) {
-                if constexpr (Util::is_among_v<U, size_t>) {
-                    out << "[" << acc << "]";
-                } else {
-                    static_assert(Util::is_among_v<U, SharedString>);
-                    out << "." << *acc;
-                }
-            },
-            x);
-    }
-    out << to_string(spec.conversion);
-    if (spec.fill || spec.conversion != FormatSpec::Conversion::str || spec.align != FormatSpec::Align::none ||
-        spec.sign != FormatSpec::Sign::minus || spec.alternate_form || spec.width > 0 ||
-        spec.grouping != FormatSpec::Grouping::none || spec.type != FormatSpec::Type::string) {
-        out << ":";
-        if (spec.fill) {
-            out << *spec.fill;
-        }
-        out << to_string(spec.align) << to_string(spec.sign);
-        if (spec.alternate_form) {
-            out << "#";
-        }
-        if (spec.width > 0) {
-            out << std::to_string(spec.width);
-        }
-        out << to_string(spec.grouping) << to_string(spec.type);
-    }
-    return out;
-}
-
 template <class T> auto print_field(T &out, FormatField const &field) -> T & {
     out << "{" << field.lhs() << field.rhs() << "}";
     return out;
@@ -933,14 +809,6 @@ auto operator<<(Util::OutputBuffer &out, BinaryOperator op) -> Util::OutputBuffe
 }
 
 // terms
-
-auto operator<<(std::ostream &out, FormatSpec const &spec) -> std::ostream & {
-    return print_spec(out, spec);
-}
-
-auto operator<<(Util::OutputBuffer &out, FormatSpec const &spec) -> Util::OutputBuffer & {
-    return print_spec(out, spec);
-}
 
 auto operator<<(std::ostream &out, FormatField const &field) -> std::ostream & {
     return print_field(out, field);
